added trip-signup functionality, enabled organizers to add/edit/delete events

+++ src/routes/events/events-route-handlers.js
@@ -3,7 +3,7 @@
  * @module "events-route-handlers"
  * @description Callback functions for Event routes
  */
-const { Event } = require('../../orm/index')
+const { Event, Trip } = require('../../orm/index')
 
 /**
 * Add Event
@@ -11,21 +11,35 @@ const { Event } = require('../../orm/index')
 * @param {*} res 
 * @param {*} next 
 */
+
+// http post :3001/events/1 organizer_user_id=1 name='Go to beach with hank' start_day='2020-1-10' end_day='2020-1-15'
 async function addEvent(req, res, next) {
-  await Event.create({
-        name: req.body.name,
-        start_day: req.body.start_day,
-        end_day: req.body.end_day,
-        trip_id: req.params.trip_id
+  const isCoordinator = await Trip.findAll({
+    where: {
+      organizer_user_id: req.body.organizer_user_id,
+      id: req.params.trip_id,
+    }
   })
-  .then(function (event, created) {
-    // console.log(event.get({
-    //     plain: true
-    // }))
-    if(created) console.log('created', created)
-    res.status(201).json(event)
-  })
-  .catch(next)
+
+  if (isCoordinator.length > 0) {
+
+    await Event.create({
+      name: req.body.name,
+      start_day: req.body.start_day,
+      end_day: req.body.end_day,
+      trip_id: req.params.trip_id
+    })
+      .then(function (event, created) {
+        // console.log(event.get({
+        //     plain: true
+        // }))
+        if (created) console.log('created', created)
+        res.status(201).json(event)
+      })
+      .catch(next)
+  } else {
+    res.status(401).json('You do not have the proper permissions to add an event.')
+  }
 }
 
 /**
@@ -35,15 +49,16 @@ async function addEvent(req, res, next) {
  * @param {*} next 
  */
 async function getEventsFromTrip(req, res, next) {
-  console.log('req.body',req.body)
-  console.log('req.params',req.params)
+  console.log('req.body', req.body)
+  console.log('req.params', req.params)
   await Event.findAll({
-    where: {trip_id: req.params.trip_id}
+    where: { trip_id: req.params.trip_id }
   })
-  .then(function (result) {
-    console.log('getEventfromTrip',result)
-    res.status(200).json(result)})
-  .catch(next)
+    .then(function (result) {
+      console.log('getEventfromTrip', result)
+      res.status(200).json(result)
+    })
+    .catch(next)
 }
 
 /**
@@ -52,17 +67,33 @@ async function getEventsFromTrip(req, res, next) {
  * @param {*} res 
  * @param {*} next 
  */
+
+
+// http PUT :3000/events/23 organizer_user_id=<organizer_user_id> name=<name-updating-to>
 async function updateEvent(req, res, next) {
-  await Event.update({
-    name: req.body.name,
-    start_day: req.body.start_day,
-    end_day: req.body.end_day,
-    },
-    {where: {id: req.params.id}}
-  )
-  .then(function (result) {
-    res.status(201).json(result)
+
+  // checks that the user at this route is the organizer of the trip at the id in params
+  const isCoordinator = await Trip.findAll({
+    where: {
+      organizer_user_id: req.body.organizer_user_id,
+      id: req.params.id,
+    }
   })
+
+  if (isCoordinator.length > 0) {
+    await Event.update({
+      name: req.body.name,
+      start_day: req.body.start_day,
+      end_day: req.body.end_day,
+    },
+      { where: { id: req.params.id } }
+    )
+      .then(function (result) {
+        res.status(201).json(result)
+      })
+  } else {
+    res.status(401).json('Access Denied')
+  }
 }
 
 /**
@@ -71,14 +102,31 @@ async function updateEvent(req, res, next) {
  * @param {*} res 
  * @param {*} next 
  */
-async function deleteEvent (req, res, next) {
-  await Event.destroy({
-    where: {id: req.params.id}
+
+
+// http delete :3001/events/17 organizer_user_id=<organizer_user_id> trip_id
+async function deleteEvent(req, res, next) {
+  console.log('req param id', req.params.id)
+
+  const isCoordinator = await Trip.findAll({
+    where: {
+      organizer_user_id: req.body.organizer_user_id,
+      id: req.body.trip_id
+    }
   })
-  .then(function (result) {
-    res.status(201).json(result)
-  })
-  .catch(next)
+
+  if (isCoordinator.length > 0) {
+    await Event.destroy({
+      where: { id: req.params.id }
+    })
+      .then(function (result) {
+        console.log('Event Deleted')
+        res.status(204).json(result)
+      })
+      .catch(next)
+  } else {
+    res.status(401).json('Access Denied')
+  }
 }
 
-module.exports = { addEvent,getEventsFromTrip,updateEvent,deleteEvent };
+module.exports = { addEvent, getEventsFromTrip, updateEvent, deleteEvent };
